+++ src/sub-clients/videos/model/UpdateVideoRequest.ts
@@ -1,0 +1,5 @@
+export interface UpdateVideoRequest {
+  title?: string;
+  description?: string;
+  promoted?: boolean;
+}